chore: update config files to use helper function for linking

+++ git/install.ps1
@@ -1,4 +1,6 @@
-Templatize "$PSScriptRoot\gitconfig.tmpl" | Set-Content "$Env:USERPROFILE\.gitconfig" -Force
-Copy-Item "$PSScriptRoot\gitignore" "$Env:USERPROFILE\.gitignore" -Force
-git config --global core.sshCommand "C:/Windows/System32/OpenSSH/ssh.exe"
-git config --global gpg.program "C:/Program Files (x86)/GnuPG/bin/gpg.exe"
+if (!$DRY_RUN) {
+	Templatize "$PSScriptRoot\gitconfig.tmpl" | Set-Content "$Env:USERPROFILE\.gitconfig" -Force
+	git config --global core.sshCommand "C:/Windows/System32/OpenSSH/ssh.exe"
+	git config --global gpg.program "C:/Program Files (x86)/GnuPG/bin/gpg.exe"
+}
+Link-File "$PSScriptRoot\gitignore" "$Env:USERPROFILE\.gitignore"
